feat: make the buttons active during game
