+++ src/modules/sprint/scss/styles.scss
@@ -8,10 +8,10 @@
     opacity: 0.5;
 }
 #game-modal-dialog {
-    margin-top: 35vh;
+    margin-top: 20vh;
 }
 #message-modal-dialog {
-    margin-top: 27vh;
+    margin-top: 20vh;
 }
 .parrots-card {
     width: 350px;
