+++ src/modules/sprint/view/game-modal.ts
@@ -1,27 +1,9 @@
-import { Router } from '../../../types/router-types';
-import { renderPage } from '../../router/services/router';
-
-export const removeModal = (modal: HTMLElement) => {
-    const close = document.querySelector('.btn-close');
-    const closeButtons = [close] as HTMLButtonElement[];
-    closeButtons.map((close: HTMLButtonElement) => {
-        close.addEventListener('click', () => {
-            modal.remove();
-            localStorage.setItem('router', Router.MAIN);
-            renderPage(Router.MAIN);
-        });
-    });
-    const modalDialog = document.querySelector('.modal-dialog');
-    modalDialog.addEventListener('click', (e) => {
-        e.stopPropagation();
-    });
-};
-
 export const gameModal = () => {
     const modal = document.createElement('div');
     modal.className = 'modal';
     modal.style.display = 'flex';
-    modal.style.backgroundColor = 'rgba(100,100,100,0.3)';
+    modal.style.backgroundColor = 'rgba(100,100,100,0)';
+    modal.style.marginTop = '200px';
     modal.innerHTML = `
     <div class="modal-dialog" id="game-modal-dialog">
     <div class="parrots-card"></div>
